Reorganised the folder structure of the project
