+++ package.json
@@ -3,10 +3,11 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --proxy-config ./src/assets/proxy.conf.dev.json",
+    "start": "ng serve --proxy-config ./src/app/configs/proxy.conf.dev.json",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
-    "test": "ng test"
+    "test": "ng test",
+    "lint": "ng lint"
   },
   "private": true,
   "dependencies": {
@@ -26,9 +27,17 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^16.2.0",
+    "@angular-eslint/builder": "16.1.2",
+    "@angular-eslint/eslint-plugin": "16.1.2",
+    "@angular-eslint/eslint-plugin-template": "16.1.2",
+    "@angular-eslint/schematics": "16.1.2",
+    "@angular-eslint/template-parser": "16.1.2",
     "@angular/cli": "~16.2.0",
     "@angular/compiler-cli": "^16.2.0",
     "@types/jasmine": "~4.3.0",
+    "@typescript-eslint/eslint-plugin": "5.62.0",
+    "@typescript-eslint/parser": "5.62.0",
+    "eslint": "^8.47.0",
     "jasmine-core": "~4.6.0",
     "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.2.0",
