+++ src/app/app.component.ts
@@ -1,6 +1,4 @@
-import { Component, OnInit } from '@angular/core';
-import { OAuthService } from 'angular-oauth2-oidc';
-import { authConfig } from 'keycloak.config';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-root',
@@ -9,31 +7,4 @@ import { authConfig } from 'keycloak.config';
 })
 export class AppComponent {
   title = 'aiod-admin-panel';
-  token: string = '';
-
-  constructor(private oauthService: OAuthService) {
-    this.configure();
-  }
-
-  private configure(): void {
-    this.oauthService.configure(authConfig);
-    this.oauthService.loadDiscoveryDocumentAndTryLogin();
-  }
-
-  public get user(): string | null {
-    if (!this.oauthService.hasValidAccessToken()) {
-      return 'Login to proceed...';
-    }
-    let claims = this.oauthService.getIdentityClaims();
-    if (!claims) return null;
-    return claims['given_name'];
-  }
-
-  public login(): void {
-    this.oauthService.initLoginFlow();
-  }
-
-  public logout(): void {
-    this.oauthService.logOut();
-  }
 }
