+++ src/app/pages/home/components/panel/panel.component.html
@@ -16,9 +16,9 @@
     <th mat-header-cell *matHeaderCellDef>Permission</th>
     <td mat-cell *matCellDef="let element">
       {{
-        element.is_accessible_for_free == true
+        element.is_accessible_for_free === true
           ? "Federated"
-          : element.is_accessible_for_free == false
+          : element.is_accessible_for_free === false
           ? "Local"
           : "NA"
       }}
